miscellaneous cleanups

+++ cpp/TooManyCooks/nqueens.cpp
@@ -43,7 +43,7 @@ template <size_t N> tmc::task<int> nqueens(int xMax, std::array<char, N> buf) {
     co_return 1;
   }
 
-  int taskCount = 0;
+  size_t taskCount = 0;
   auto tasks = std::ranges::views::iota(0UL, N) |
                std::ranges::views::filter([xMax, &buf](int y) {
                  buf[xMax] = y;
