+++ cpp/tbb/nqueens.cpp
@@ -41,7 +41,7 @@ template <size_t N> void nqueens(int xMax, std::array<char, N> buf, int& out) {
     return;
   }
 
-  int taskCount = 0;
+  size_t taskCount = 0;
   std::array<int, nqueens_work> results;
   auto tasks =
     std::ranges::views::iota(0UL, N) |
